fix: hydra on cindy(aarch64-linux) seems to hang with nix 2.11.0
A build step is performed but never stops, on the hydra host machine
found a process `nix-store --builders --serve --write` that waits
indefinitely, `strace` shows the process is reading from a pipe but
could not get any data, on the pipe's write end there is another process
that polls for something (did not look into this) and reads from a
socket, `lsof` then shows the socket's read and write ends are both from
the second process.  After reverting to nix 2.10.3 the hanging problem
is gone.

Possibly related:
  - <https://github.com/NixOS/hydra/issues/366>
  - <https://github.com/NixOS/nix/issues/2560>
  - <https://github.com/NixOS/nix/issues/2260>

Revert "chore: bump flake inputs"

This reverts commit 1d75b21e8b44fb7d4f558da193a30dd6dc974fe3.

Signed-off-by: Gaoyang Zhang <gy@blurgy.xyz>