zjuwlan-login-script: ensure checkbox is checked
The ZJUWLAN login website has been updated with the checkbox being
renamed to something like "I agree those terms" and by default checked.
The previous zjuwlan-login-script toggles it to unchecked thus failing
to submit the login form.  This commit ensures the checkbox is always
checked instead of always assuming the checkbox is unchecked then
toggling it.

Signed-off-by: Gaoyang Zhang <gy@blurgy.xyz>